+++ src/forms/form-error/form-error.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default ({error, clearError}) => {
+  return (
+    error && <div className='FormError'>{error}</div> || null
+  );
+};
