+++ src/forms/form/form.js
@@ -2,7 +2,6 @@ import { connect } from 'react-view-models';
 import DefineMap from 'can-define/map/';
 import {devWarning} from 'auth-component/utils';
 import View from './form.jsx';
-import './forms.less';
 
 export const ViewModel = DefineMap.extend({
   '*': {
@@ -22,11 +21,29 @@ export const ViewModel = DefineMap.extend({
   Model: 'any',
 
   /**
+   * `service` is a FeathersJS service. Its create method will be used to submit
+   * data to the server.
+   */
+  service: 'any',
+
+  /**
    * There are a few warnings that will show up by default. They can be turned
    * off by setting `suppressWarnings` to true.
    */
   suppressWarnings: {
     value: false
+  },
+
+  /**
+   * When an error is returned from the server, it will end up here.
+   */
+  error: 'string',
+
+  /**
+   * Clears the error message.
+   */
+  clearError () {
+    this.error = undefined;
   },
 
   /**
@@ -41,15 +58,16 @@ export const ViewModel = DefineMap.extend({
   },
 
   /**
-   * `formSubmitted` is the handler for the form submit. It calls `onSubmit`
+   * `formSubmitted` is the handler for the form. It calls `onSubmit`
    * with the auth data.
    */
   formSubmitted (values) {
     if (this.strategy) {
       values.strategy = this.strategy;
     }
+    this.clearError();
     this.onSubmit(values)
-      .then(response => this.onSubmitSuccess(response))
+      .then(response => this.onSuccess(response))
       .catch(error => this.uiError(error));
   },
 
@@ -71,26 +89,27 @@ export const ViewModel = DefineMap.extend({
   },
 
   /**
-   * `onSubmitSuccess` function gets run when a successful onSubmit response was received.
+   * `onSuccess` function gets run when a successful onSubmit response was received.
    * In most cases, it will need to be overwritten to handle custom requirements.
    */
-  onSubmitSuccess (data) {
-    this.warn(`Pass an "onSubmitSuccess" function to the ${this.formName} to handle success.`);
+  onSuccess (data) {
+    this.warn(`Pass an "onSuccess" function to the ${this.formName} to handle success.`);
   },
 
   /**
    * `uiError` makes sure the UI responds properly to any error received.
-   * It calls `onSubmitError`.
+   * It calls `onError`.
    */
   uiError (error) {
-    this.onSubmitError(error);
+    this.error = error.message || error;
+    this.onError(error);
   },
 
   /**
-   * When submit fails, the `onSubmitError` callback can be used to handle custom
+   * When submit fails, the `onError` callback can be used to handle custom
    * logic in your app.
    */
-  onSubmitError (error) {
+  onError (error) {
     this.warn(error);
   }
 });
