+++ src/forms/form/test.html
@@ -1,3 +1,3 @@
 <script src="../../../node_modules/steal/steal.js"
   mocha="bdd"
-  main="auth-component/forms/input/input_test"></script>
+  main="auth-component/forms/form/form_test"></script>
