--- src/forms/input/input.less
@@ -1,20 +1,0 @@
-form.auth-component-form {  
-  input {
-    border-bottom: solid 1px #E6E6E6;
-    border-top: none;
-    border-left: none;
-    border-right: none;
-    height: 30px;
-    width: calc(100% - 10px);
-    text-align: center;
-    padding: 6px 0 2px;
-    margin-bottom: 4px;
-    outline: none;
-    font-size: 14px;
-    box-sizing: content-box;
-
-    &:focus {
-      border-bottom: solid 1px #77BFED;
-    }
-  }
-}
