+++ src/forms/local-login/demo.html
@@ -19,7 +19,7 @@
       }
     </style>
     <section root="true"></section>
-    <script src="/node_modules/steal/steal.js" main="auth-component/forms/login/login-demo"></script>    
+    <script src="/node_modules/steal/steal.js" main="auth-component/forms/local-login/demo"></script>    
   </body>
 </html>
