--- src/forms/local-login/login.md
@@ -1,3 +1,0 @@
-# Email / Password Login Form
-
-The
