+++ src/forms/local-signup/demo.html
@@ -19,7 +19,7 @@
       }
     </style>
     <section root="true"></section>
-    <script src="/node_modules/steal/steal.js" main="auth-component/forms/signup/signup-demo"></script>    
+    <script src="../../../node_modules/steal/steal.js" main="auth-component/forms/local-signup/demo"></script>    
   </body>
 </html>
