--- src/forms/test.html
@@ -1,3 +1,0 @@
-<script src="../../node_modules/steal/steal.js"
-  mocha="bdd"
-  main="auth-component/forms/forms_test"></script>
